requirements.txt
opencv-python
numpy
Pillow
imagehash
tqdm
PyYAML
prompt_toolkit
plyfile
